autobuild

+++ tests/test_helper.js
@@ -1,4 +1,6 @@
+const bcrypt = require('bcrypt')
 const Note = require('../models/note')
+const User = require('../models/user')
 
 const initialNotes = [
   {
@@ -11,12 +13,34 @@ const initialNotes = [
   },
 ]
 
-const nonExistingId = async () => {
-  const note = new Note({ content: 'will_be_deleted' })
-  await note.save()
-  await note.deleteOne()
+let _noteId = undefined
+let _userId = undefined
 
-  return note._id.toString()
+const nonExistingNoteId = async () => {
+  if (!_noteId) {
+    const note = new Note({ content: 'null content' })
+    await note.save()
+    await note.deleteOne()
+
+    _noteId = note._id.toString()
+  }
+  return _noteId
+}
+
+const existingUserId = async () => {
+  if (!_userId) {
+    const saltRounds = 10
+    const passwordHash = await bcrypt.hash('null', saltRounds)
+
+    const user = await User.create({
+      username: 'null',
+      name: 'null',
+      passwordHash,
+    })
+
+    _userId = user._id.toString()
+  }
+  return _userId
 }
 
 const notesInDb = async () => {
@@ -29,9 +53,16 @@ const noteInDb = async (id) => {
   return note.toJSON()
 }
 
+const usersInDb = async () => {
+  const users = await User.find({})
+  return users.map((user) => user.toJSON())
+}
+
 module.exports = {
   initialNotes,
-  nonExistingId,
+  nonExistingNoteId,
+  existingUserId,
   notesInDb,
   noteInDb,
+  usersInDb,
 }
